A lot of new stuff, too lazy to describe all this

+++ src/main/java/net/lopymine/ip/element/base/IHoverableElement.java
@@ -8,7 +8,7 @@ public interface IHoverableElement {
 
 	boolean isHovered();
 
-	default void updateHovered(InventoryCursor cursor, int x, int y, int width, int height) {
+	default void updateHovered(InventoryCursor cursor, float x, float y, float width, float height) {
 		boolean bl = cursor.getMouseX() > x && cursor.getMouseX() < x + width && cursor.getMouseY() > y && cursor.getMouseY() < y + height;
 		this.setHovered(bl);
 	}
